refactor: add StatementWrongDateSelected exception and handler for date mismatches

+++ backend/services/statement_service.py
@@ -10,25 +10,30 @@ from datetime import date
 from mappers.transaction_mapper import TransactionMapper
 from utils import utils
 from domain.transaction_builder import TransactionBuilder
+from exceptions.domain import StatementWrongDateSelected
 
 class StatementService:
   def __init__(self, db: Session):
     self.db = db
     self.repository = StatementRepository(db)
 
-  def generate_monthly_statement(self, file: bytes, date: date) -> StatementDTO:
+  def generate_monthly_statement(self, file: bytes, user_selected_date: date) -> StatementDTO:
     table = extract_table_from_pdf(file)
     df = normalize_statement_dataframe(build_statement_dataframe(table))
     
     transactions = TransactionMapper.from_df(df)
+    statement_date = transactions[0].transaction_date
     
-    record = self.repository.get_statement_via_date(date)
+    if (statement_date.year, statement_date.month) != (user_selected_date.year, user_selected_date.month):
+      raise StatementWrongDateSelected(user_selected_date=user_selected_date, statement_date=statement_date)
+    
+    record = self.repository.get_statement_via_date(user_selected_date)
     if record:
       self.repository.delete_statement(record)
     
-    self.repository.create_statement(transactions, date)
+    self.repository.create_statement(transactions, user_selected_date)
     
-    return self.get_monthly_statement(date)
+    return self.get_monthly_statement(user_selected_date)
   
   def get_monthly_statement(self, date: date) -> StatementDTO:
     end_date = utils.get_end_of_month(date)
